u16

+++ jul1/src/combinators/forbid_follows.rs
@@ -3,12 +3,12 @@ use std::rc::Rc;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Default)]
 pub struct ForbidFollowsData {
-    pub prev_match_ids: u32, // Using a bitset
+    pub prev_match_ids: u16, // Using a bitset
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ForbidFollows {
-    pub(crate) match_ids: u32, // Using a bitset
+    pub(crate) match_ids: u16, // Using a bitset
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
@@ -16,7 +16,7 @@ pub struct ForbidFollowsClear {}
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ForbidFollowsCheckNot {
-    pub(crate) match_ids: u32,
+    pub(crate) match_ids: u16,
 }
 
 impl CombinatorTrait for ForbidFollows {
@@ -47,7 +47,7 @@ impl CombinatorTrait for ForbidFollowsCheckNot {
 pub fn forbid_follows(match_ids: &[usize]) -> ForbidFollows { // Using a bitset
     let mut bitset = 0;
     for &id in match_ids {
-        bitset |= 1 << TryInto::<u32>::try_into(id).unwrap();
+        bitset |= 1 << TryInto::<u16>::try_into(id).unwrap();
     }
     ForbidFollows { match_ids: bitset }
 }
